Rework the pbuf code as a set of inline functions.
This was primarily intended to make it possible to use pbufs from interrupt context.  It may increase the cost of handling them outside, however, so it may end up being the case that we need two sets of functions.

Also simplify the pbuf 'pointer' handling so that locating the 'next' buffer in a chain doesn't involve a pointer deref.