changed packet transmit completion strategy
the PKSENT interrupt was sometimes not triggering on successful
send. This patch changes the strategy to use the chip power status,
and watches for the chip to enter and then exit transmit mode. That
turns out to be a much more reliable way of detecting transmit
completion, and prevents us getting timeouts in the transmit code