latency improvements POC - described below.
These improvements should be wire compatible/not change behavior unless you
change the new params from their defaults.

1) Allow shrinking the max transmit window.  Use ATS15=1023 for now
FIXME: I need to change the units of this to something like usecs or bytes...

2) Ensure only one rc-override (the freshest) is sent in each transmit window
(usually - for edge cases where we are out of sync with mavlink a second
could be sent).  It is worth noting that the benefit from this change is
very small (in fact, I'm still working on a test harness that can measure it)
because Andropilot is careful to avoid sending too many rc-overrides or
overrides with stale data.
To enable ATS6=2  (ATS6=1 is normal behavior)

There are a few FIXMEs still - so DO NOT MERGE.  In particular:
I shrunk the rx tx buffer to 256 bytes to find some more space.  Tridge,
did you mention something about we could shrink the rx buffer instead?
Possibly by using a buffer length rather than anding with a mask.