Reduced storage needed for golay code. (from RFDesign/SiK fork)
The high bite of the golay_encode23 table was always the index/32 so
there’s no real need to store it, reducing the table size by half.

The same should be doable with two tables for decode, one of uint16_t
and one of uint8_t to reduce size by a quarter. This might not be
beneficial though, since it’ll mean two memory accesses, and the encode
is helped by being able to do a shift of v, but no such (easily
calculable) optimisation exists in the decode code.

Currently doesn’t compile with RFDesign's code (I assume because ECC was
removed to free up space for other features), but hopefully
it’s a useful improvement saving some room for other data and hopefully
speeding up the encoding.

Conflicts:
	Firmware/radio/golay23.h