atcommander: initial commit
Added a new file that contains the ATCommandSet class, which wraps
the AT command set on the radio. Currently, it can list out device
information, change parameters (settings), and reboot both the locally
attached and remotely linked radios. It is moderately tested, and
lacks implementation of some AT commands.

This file also includes a command line interface for displaying and
changing parameters. It tries its best to keep the user out of trouble.